CreateLog and Real Time Event Service

+++ Oog.API/Common/DTOs/AccountDto.cs
@@ -5,7 +5,7 @@ namespace API.Common.DTOs;
 
 public class AccountDto
 {
-   public required long Id { get; set; }
+   public required int Id { get; set; }
    public required string Username { get; set; }
    public required DateTime CreatedAt { get; set; }
    public required DateTime UpdatedAt { get; set; }
