+++ Oog.API/Common/DTOs/AppDto.cs
@@ -2,8 +2,8 @@
 
 public class AppDto
 {
-   public required long Id { get; set; }
-   public required long EnvId { get; set; }
+   public required int Id { get; set; }
+   public required int EnvId { get; set; }
    public required string Name { get; set; }
    public required string Passkey { get; set; }
 }
