+++ Oog.API/Common/DTOs/EnvironmentDto.cs
@@ -2,7 +2,7 @@
 
 public class EnvironmentDto
 {
-    public required long Id { get; set; }
+    public required int Id { get; set; }
     public required string Name { get; set; }
-    public required long OwnerId { get; set; }
+    public required int OwnerId { get; set; }
 }
